spark/requirements.txt
# pyspark==3.5.1
venv-pack
pyyaml
ua-parser
user-agents
psycopg2-binary
# confluent_kafka
# apache-airflow-providers-apache-kafka